runtime-core: Move ArgumentRegisterAllocator into `state`.
